Abstract out guid item view
